refactor: Phase 5 slice 5b.3 — extract ChronologyPanel + FriendsPanel (#118)
First component decomposition of AnimeDetailView. Stacked on slice 5b.2 (#122).

Two low-risk sidebar panels move into components/detail/:
- ChronologyPanel.vue (311 L) — Shikimori related-anime list. Props:
  shikiRelated + relatedLoading; v-model:collapsed. Owns KIND_LABELS,
  STATUS_LABELS, the click/keyboard navigation, scoped CSS (.related-* +
  .status-*). Calls useLibraryStore().openAnime() directly.
- FriendsPanel.vue (217 L) — Friends' Shikimori watch status grid. Props:
  friendsRates + friendsLoading + numberOfEpisodes; v-model:collapsed. Owns
  the friendsSummary computed (status counts) + STATUS_LABELS + scoped CSS.

AnimeDetailView.vue: 2898 → 2441 lines (~460-line shrink). The original
markup is replaced by two single-line component invocations with v-model
for collapsed state. KIND_LABELS, STATUS_LABELS, and friendsSummary are
no longer needed at the parent level and are deleted. The .related-* +
.friends-* + .friend-* + .status-* scoped CSS rules (~270 lines) move
out of the parent and into the respective component scoped blocks.

Behavior byte-identical:
- Same v-if gates (anime.myAnimeListId && (relatedLoading ||
  shikiRelated.length > 0)) and (anime.myAnimeListId && (shikiUser ||
  !shikiUserChecked)).
- Same collapsed-state toggle behavior via v-model.
- Same click/keyboard navigation in chronology (clickable rows for
  available smotret entries, disabled for current/unavailable).
- Same friend list grid layout + status badges.
- Status-* color classes preserved per-component since each tab scopes
  its styles independently.

Skip Detection panel and Shikimori panel stay in AnimeDetailView — they
need their own composables before clean component extraction (skip
detection has 3 broadcast subscriptions; Shikimori panel has the rate
edit form + offline indicator + sync state).

Quality gate: typecheck + lint (0 errors, 7 pre-existing warnings) +
format:check + 171 tests + build + grep gate all green.

Stacked PR: base = phase5-slice-5b2-downloads-composable. Will be
retargeted to main when its base merges (per
[[feedback_stacked_pr_base_branch]]).

Cumulative Phase 5 slice 5b trajectory:
- AnimeDetailView.vue: 3624 (main) → 3260 (5b.1) → 2898 (5b.2) → 2441 (5b.3)
- ~33% script reduction across 3 PRs.
- New testable surface: 4 composables (1063 LoC) + 60 unit tests (1072 LoC tests).
- 2 sub-components extracted (528 LoC).

Part of #84 (Phase 5 slice 5b.3).
